+++ pkg/storage/endpoint/min_resolved_ts.go
@@ -18,7 +18,7 @@ import (
 	"strconv"
 
 	"github.com/tikv/pd/pkg/errs"
-	"github.com/tikv/pd/pkg/utils/etcdutil/keypath"
+	"github.com/tikv/pd/pkg/utils/keypath"
 )
 
 // MinResolvedTSPoint is the min resolved ts for a store
