+++ pkg/storage/endpoint/replication_status.go
@@ -18,7 +18,7 @@ import (
 	"encoding/json"
 
 	"github.com/tikv/pd/pkg/errs"
-	"github.com/tikv/pd/pkg/utils/etcdutil/keypath"
+	"github.com/tikv/pd/pkg/utils/keypath"
 )
 
 // ReplicationStatusStorage defines the storage operations on the replication status.
