+++ pkg/storage/storage_gc_test.go
@@ -22,7 +22,7 @@ import (
 	"github.com/pingcap/failpoint"
 	"github.com/stretchr/testify/require"
 	"github.com/tikv/pd/pkg/storage/endpoint"
-	"github.com/tikv/pd/pkg/utils/etcdutil/keypath"
+	"github.com/tikv/pd/pkg/utils/keypath"
 )
 
 func testGCSafePoints() []*endpoint.GCSafePointV2 {
