+++ tools/pd-backup/pdbackup/backup.go
@@ -26,7 +26,7 @@ import (
 	"strconv"
 
 	"github.com/tikv/pd/pkg/utils/etcdutil"
-	"github.com/tikv/pd/pkg/utils/etcdutil/keypath"
+	"github.com/tikv/pd/pkg/utils/keypath"
 	"github.com/tikv/pd/pkg/utils/typeutil"
 	"github.com/tikv/pd/server/config"
 	clientv3 "go.etcd.io/etcd/client/v3"
